[envtest]Move the API simulator to its own file
